adicionado sweetalert

+++ ONGWeb2/src/js/editar-ong.js
@@ -69,40 +69,62 @@ function handleDrop(event) {
 }
 
 function deleteOng() {
-    const token = localStorage.getItem('token');
-    axios.delete('http://localhost:3000/api/ong/', { 
-        headers: {
-            'Authorization': `Bearer ${token}`
+    Swal.fire({
+        title: "Tem certeza?",
+        text: "Sua ONG será excluída permanentemente do sistema !",
+        icon: "warning",
+        showCancelButton: true,
+        confirmButtonColor: "#3085d6",
+        cancelButtonColor: "#d33",
+        confirmButtonText: "Sim, quero excluir",
+        cancelButtonText: 'Cancelar'
+      }).then((result) => {
+        if (result.isConfirmed) {
+            const token = localStorage.getItem('token');
+            axios.delete('http://localhost:3000/api/ong/', { 
+                headers: {
+                    'Authorization': `Bearer ${token}`
+                }
+            })
+            .then(response => {
+                localStorage.removeItem('token')
+                console.log('Response:', response.data);
+                Swal.fire({
+                    title: "Sucesso!",
+                    text: "Sua ONG foi removida do sistema.",
+                    icon: "success",
+                }).then((result) => {
+                    window.location.href = "/ONGWeb2/src/ui/";
+                })
+            })
+            .catch(error => {
+              console.error('Error:', error);
+            });
         }
-    })
-    .then(response => {
-      console.log('Response:', response.data);
-    })
-    .catch(error => {
-      console.error('Error:', error);
-    });
+      });
 }
 
+function validateForm(form) {
+    return new Promise((resolve, reject) => {
+        console.log(form)
+        
+        if (form.get('senha') !== form.get('confirmarSenha')) 
+            reject('As senhas não coincidem!')
 
-function uploadImageTest() {
+        if (!form.get('senha')) 
+            reject('A senha deve ser informada!')
 
-    const formData = new FormData();
-    const images = document.getElementById('imagemInput').files;
-    const token = localStorage.getItem('token');
-    
-    formData.append('image', images[0]);
-    axios.post('http://localhost:3000/api/ong/images', formData, {
-        headers: {
-            'Content-Type': 'multipart/form-data',
-            'Authorization': `Bearer ${token}`
-        }
+        if (!form.get('email'))
+            reject('O E-mail deve ser informado!')
+
+        if (!form.get('cnpj')) 
+            reject('O CNPJ deve ser informado!')
+
+        if (!form.get('image')) 
+            reject('Insira uma imagem para a ONG!')
+
+        resolve();
     })
-    .then(response => {
-      console.log('Response:', response.data);
-    })
-    .catch(error => {
-      console.error('Error:', error);
-    });
 }
 
 async function update() {
@@ -120,24 +142,33 @@ async function update() {
     formData.append('numero', document.getElementById('numero').value),
     formData.append('cidade', document.getElementById('cidade').value)
 
-    // for (const file of document.getElementById('imagemInput').files) {
-    //     const base64String = await imageToBase64(file);
-    //     obj.imagens.push(base64String);
-    // }
+    if (formData.get('senha') !== formData.get('confirmarSenha')) { 
+        Swal.fire({
+            backdrop: false,
+            width: '200px',
+            position: "top-end",
+            icon: "error",
+            text: "As senhas não coincidem!",
+            showConfirmButton: false,
+            timer: 1500
+        });
+        return
+     } 
 
-    // if (obj.senha !== obj.confirmarSenha) { return alert("Confirmação de Senha Inválida"); } 
     const token = localStorage.getItem('token');
     
-    axios.put('http://localhost:3000/api/ong/update', formData, { 
-        headers: {
-            'Content-Type': 'application/json',
-            'Authorization': `Bearer ${token}`
-        }
-    })
-    .then(response => {
-      console.log('Response:', response.data);
-    })
-    .catch(error => {
-      console.error('Error:', error);
-    });
+    validateForm(formData).then(() => {
+        axios.put('http://localhost:3000/api/ong/update', formData, { 
+            headers: {
+                'Content-Type': 'application/json',
+                'Authorization': `Bearer ${token}`
+            }
+        })
+        .then(response => {
+            console.log('Response:', response.data);
+        })
+        .catch(error => {
+          console.error('Error:', error);
+        });
+    }).catch(error => fireErrorMessage(error))
 }
